refactor(t746): move autosaveConstants to lib/ per convention
Constants/types file belongs in src/lib/, not src/hooks/ which is
reserved for use* hook files. Update import paths in both hooks.
